+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 # Revision history for delivery-status-notification
 
-## 0.1.1.0 -- 2024-08-18
+## 0.2.0.0 -- 2024-08-19
 
-* added attoparsec `Email.StatusCode.parser`
-* added `Email.DSN`
+* added attoparsec status code parser
+* added `Email.DSN.DeliveryStatus`
+* moved `Email.StatusCode` to `Email.DSN.StatusCode`
 
 ## 0.1.0.0 -- 2024-08-18